Moved road logic around and disabled it for now. Added positions to simple game obj. Removed Id from barb tribe and stronghold. Should use ObjectId instead. Cleaned up simple game obj creation to always pass in ObjectId. Other refactorings.

+++ Game/Data/Tribe/TribeManager.cs
@@ -49,7 +49,7 @@ namespace Game.Data.Tribe
 
         public void Add(ITribe tribe)
         {
-            tribe.Id = (uint)tribeIdGen.GetNext();
+            tribe.Id = tribeIdGen.GetNext();
             if (!Tribes.TryAdd(tribe.Id, tribe))
             {
                 return;
